Se cambiaron los datos a Float y se permite ingresar el 0

+++ TP1 Calculadora/funciones.h
@@ -5,12 +5,12 @@
 
 #endif // FUNCIONES_H_INCLUDED
 
-int showMenu(int x, int y, int flag1, int flag2);
+int showMenu(float x, float y, int flag1, int flag2);
 int getOption(int min, int max, char message[], char messageError[]);
-int getNumber(char message[], char messageError[]);
-int add(int x, int y);
-int substract(int x, int y);
-int multiplicate(int x, int y);
-float divide(int x, int y);
-long long int factorial(int num);
-void showResults(int x, int y, int addition, int substraction, int multiplication, float division, long long int fact1, long long int fact2);
+float getNumber(char message[], char messageError[]);
+float add(float x, float y);
+float substract(float x, float y);
+float multiplicate(float x, float y);
+float divide(float x, float y);
+long long int factorial(float num);
+void showResults(float x, float y, float addition, float substraction, float multiplication, float division, long long int fact1, long long int fact2);
